requirements.txt
requests==2.31.0
urllib3==2.0.7
python-dotenv==1.0.0
colorama==0.4.6
tabulate==0.9.0
streamlit==1.28.0
plotly==5.17.0
pandas==2.1.0
